Fix cyclic dependency issue
