+++ src/gb/models/glucoseModel.ts
@@ -14,3 +14,19 @@ export enum GlucoseUnit {
     MMOL_L = 'mmol/L',
     MG_DL = 'mg/dL'
 }
+
+
+/** //TODO move
+ * Different record type meanings
+ * Glucose levels (0 & 1) are in mmol/L
+ * Insulin (4) includes both rapid-acting insulin and long-acting insulin (in units)
+ * Carbohydrates are in grams
+ */
+ export enum RecordType {
+    HISTORIC_GLUCOSE_LEVEL = 0,
+    SCAN_GLUCOSE_LEVEL = 1,
+    STRIP_GLUCOSE_LEVEL = 2,
+    INSULIN = 4,
+    CARBOHYDRATES = 5,
+    NOTES = 6
+}
