+++ src/services/modelParser.ts
@@ -15,7 +15,11 @@ export abstract class ModelParser {
     ) {
         this.gbClient = new GameBusClient(
             //TODO optional, because it is not used?
-            new TokenHandler(userInfo.accessToken, userInfo.refreshToken as string, userInfo.playerId)
+            new TokenHandler(
+                userInfo.accessToken,
+                userInfo.refreshToken as string,
+                userInfo.playerId
+            )
         );
     }
